doc: clarify the social contract after the XZ backdoor
In light of the XZ backdoor, I feel the need to make explicit that I’m
happy to get contributions and comments, but that folks shouldn’t expect
to get commit rights or even an answer.
This is a hobby project and they can be weaponized by nefarious actors.
OSS maintainer burn-out is also all too common, so let’s try not to
overwork ourselves.

As a side note, the criteria to give out commit rights has been my rule
all along, but it was unwritten.